# apiLayer/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Api UMbook",
        "description": "Aqui encontra-se a documentação de toda a api de dados da UMbook, incluindo todas as rotas e toda a informação à cerca do que é precisar passar como parâmetros, bem como os resultados que se podem esperar ao executar aquela rota.\n Muitas das rotas encontram-se protegidas, não se esqueça de realizar o login na aba dos utilizadores com as suas credenciais para ter acesso às rotas que lhe são permitido aceder. Receberá um token de acesso e terá que o copiar e colar no botão 'Authorize' no canto superior direito."
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3049",
    "basePath": "/api",
...
    "paths": {
        "/cursos": {
            "get": {
                "tags":[
                    "cursos"
                ],
                "summary": "Todos os ids dos cursos registados.",
                "description": "Retorna uma lista com todos os ids dos cursos presentes e respetivas designações.",
                "responses": {
                    "200": {
                        "description": "Lista de Curso",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "curso": {
                                        "type": "string"
                                    },
                                    "designacao": {
                                      "type":"string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post":{
                "tags":[
                    "cursos"
                ],
                "summary": "Inserir um novo curso.",
                "description": "Adiciona um novo curso à base de dados.",
                "parameters": [
                    {
                      "name": "id",
                      "in": "formData",
                      "description": "ID do curso",
                      "required": true,
                      "type": "string"
                    },
                    {
                      "name": "nome",
                      "in": "formData",
                      "description": "Nome do curso",
                      "required": true,
                      "type": "string"
                    }
                  ],
                "responses": {
                    "200": {
                        "description": "Retorna o id do Curso criado",
                        "schema": {
                            "type": "string",
                            "items": {
                                "properties": {
                                    "curso": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "401":{
                      "description": "Falta-lhe o token de acesso ou o token não é válido."
                    }
                },
                "security": [
                  {
                    "api_key": []
                  }
                ]
            }
        },
...
        "/anos": {
            "post":{
                "tags":[
                    "anos"
                ],
                "summary": "Inserir um novo ano de um dado curso.",
                "description": "Para inserir um novo ano a um curso, é necessário ser Admin.",
                "parameters": [
                    {
                      "name": "nome",
                      "in": "formData",
                      "description": "Nome do ano",
                      "required": true,
                      "type": "string"
                    },
                    {
                        "name": "anoLetivo",
                        "in": "formData",
                        "description": "Ano letivo do ano",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "idCurso",
                        "in": "formData",
                        "description": "ID do curso que o ano pertence",
                        "required": true,
                        "type": "string"
                      }
                  ],
                "responses": {
                    "200": {
                        "description": "Retorna o id do ano criado",
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401":{
                      "description": "Falta-lhe o token de acesso ou o token não é válido."
                    },
                    "403":{
                      "description": "Não possuí nível de acesso suficiente."
                    }
                },
                "security": [
                  {
                    "api_key": []
                  }
                ]
            }
        },
...
        "/cadeiras": {
            "post":{
                "tags":[
                    "cadeiras"
                ],
                "summary": "Inserir uma nova cadeira de um dado ano.",
                "description": "Para adicionar uma nova cadeira à base de dados, é necessário ser Admin ou Responsável.",
                "parameters": [
                    {
                      "name": "nome",
                      "in": "formData",
                      "description": "Nome da cadeira",
                      "required": true,
                      "type": "string"
                    },
                    {
                      "name": "conteudosProgramaticos",
                      "in": "formData",
                      "description": "Conteúdos programáticos da cadeira",
                      "required": true,
                      "type": "string"
                    },
                    {
                        "name": "idAno",
                        "in": "formData",
                        "description": "ID do ano que a cadeira pertence",
                        "required": true,
                        "type": "string"
                      }
                  ],
                "responses": {
                    "200": {
                        "description": "Retorna o id da cadeira criada",
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401":{
                      "description": "Falta-lhe o token de acesso ou o token não é válido."
                    },
                    "403":{
                      "description": "Não tem nível de acesso suficiente!"
                    }
                },
                "security": [
                  {
                    "api_key": []
                  }
                ]
            }
        },
...
        "/eventos": {
            "get": {
                "tags":[
                    "eventos"
                ],
                "summary": "Retorna todos os eventos registados.",
                "description": "Retorna uma lista com todos os eventos presentes.",
                "responses": {
                    "200": {
                        "description": "Retorna toda a informação relativa aos eventos que possuem uma determinada parceria.",
                        "schema": {
                            "type": "array",
                            "items":{
                                "properties": {
                                    "info": {
                                        "$ref": "#/definitions/Evento"
                                    },
                                    "participantes":{
                                        "type":"array",
                                        "items": {
                                            "type":"object",
                                            "properties": {
                                                "idUtilizador":{
                                                    "type":"string"
                                                },
                                                "nome":{
                                                    "type":"string"
                                                }
                                            }
                                        }
                                    },
                                    "parcerias":{
                                        "type": "array",
                                        "items": {
                                            "type":"object",
                                            "properties": {
                                                "idCurso":{
                                                    "type":"string"
                                                },
                                                "nome":{
                                                    "type":"string"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                      },
                      "401":{
                        "description": "Falta-lhe o token de acesso ou o token não é válido."
                      }
                  },
                  "security": [
                    {
                      "api_key": []
                    }
                  ]
            },
            "post":{
                "tags":[
                    "eventos"
                ],
                "summary": "Inserir um novo evento.",
                "description": "Para inserir um novo evento, é necessário ser Admin.",
                "parameters": [
                    {
                      "name": "nome",
                      "in": "formData",
                      "description": "Nome do evento",
                      "required": true,
                      "type": "string"
                    },
                    {
                        "name": "conteudo",
                        "in": "formData",
                        "description": "Informação adicional do evento",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "dataInicio",
                        "in": "formData",
                        "description": "Data e hora a que o evento começa",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "dataFim",
                        "in": "formData",
                        "description": "Data e hora a que o evento acaba",
                        "required": true,
                        "type": "string"
                    },
                    {
                      "name": "parcerias",
                      "in": "formData",
                      "description": "As parcerias do Evento",
                      "required": true,
                      "type": "array",
                      "items":{
                        "type": "string"
                      }
                  }
                  ],
                "responses": {
                    "200": {
                        "description": "Retorna o id do evento criado",
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401":{
                      "description": "Falta-lhe o token de acesso ou o token não é válido."
                    },
                    "403":{
                      "description": "Não possuí nível de acesso suficiente."
                    }
                },
                "security": [
                  {
                    "api_key": []
                  }
                ]
            }
        },
...
        "/publicacoes":{
            "post":{
                "tags":[
                    "publicacoes"
                ],
                "summary": "Inserir uma nova publicação.",
                "description": "Adiciona uma nova publicação à base de dados.",
                "parameters": [
                    {
                      "name": "idGrupo",
                      "in": "formData",
                      "description": "Id do grupo em que a publicação vai ser publicada",
                      "required": true,
                      "type": "string"
                    },
                    {
                        "name": "conteudo",
                        "in": "formData",
                        "description": "Conteudo da publicação",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "idUtilizador",
                        "in": "formData",
                        "description": "Id do utilizador que a publicou",
                        "required": true,
                        "type": "string"
                    }
                  ],
                "responses": {
                    "200": {
                        "description": "Retorna o id da publicação criada",
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401":{
                      "description": "Falta-lhe o token de acesso ou o token não é válido."
                    }
                },
                "security": [
                  {
                    "api_key": []
                  }
                ]
            }
        },
...
        "/utilizadores":{
          "get":{
            "tags":[
              "utilizadores"
          ],
          "summary": "Retorna todos os utilizadores da aplicação.",
          "description": "Para tal, é necessário o utilizador que faz o pedido ser Admin.",
          "responses": {
            "200": {
              "description": "Lista de Utilizadores.",
              "schema":{
                "type":"array",
                "items": {
                  "properties": {
                    "schema": {
                      "$ref": "#/definitions/Utilizador"
                    }  
                  }
                }
              }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            },
            "403":{
              "description": "Não possuí nível de acesso suficiente."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
          },
          "post": {
            "tags":[
              "utilizadores"
            ],
            "summary": "Insere um novo utilizador.",
            "description": "O utilizador é adicionado á base de dados.",
            "parameters": [
              {
                "name": "id",
                "in": "formData",
                "description": "Email do utilizador",
                "required": true,
                "type": "string"
              },
              {
                "name": "nome",
                "in": "formData",
                "description": "Nome do utilizador",
                "required": true,
                "type": "string"
              },
              {
                "name": "numeroAluno",
                "in": "formData",
                "description": "Número de aluno",
                "required": true,
                "type": "string"
              },
              {
                "name": "idCurso",
                "in": "formData",
                "description": "Identificador do curso a que pertence",
                "required": true,
                "type": "string"
              },
              {
                "name": "idAno",
                "in": "formData",
                "description": "Identificador do ano que frequenta",
                "required": true,
                "type": "string"
              },
              {
                "name": "dataNascimento",
                "in": "formData",
                "description": "Data de Nascimento",
                "required": true,
                "type": "string"
              },
              {
                "name": "sexo",
                "in": "formData",
                "description": "Género do utilizador",
                "enum": [ "Masculino", "Feminino", "Outro"],
                "required": true,
                "type": "string"
              },
              {
                "name": "numeroTelemovel",
                "in": "formData",
                "description": "Número de telemóvel",
                "required": true,
                "type": "string"
              },
              {
                "name": "password",
                "in": "formData",
                "description": "Password",
                "required": true,
                "type": "string"
              }
            ],
            "responses": {
              "200": {
                "description": "O utilizador foi inserido com sucesso."
              },
              "401":{
                "description": "Falta-lhe o token de acesso ou o token não é válido."
              },
              "403":{
                "description": "Não possuí nível de acesso suficiente."
              }
            },
            "security": [
              {
                "api_key": []
              }
            ]
          }
        },
...
    },
      "securityDefinitions":{
        "api_key":{
          "type": "apiKey",
          "in": "query",
          "name": "token"
        }
      },
      "security": [{
         "api_key": []
      }
    ],
    "definitions": {
        "Utilizador": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "nomeUtilizador": {
                "type": "string"
              },
              "numeroAluno": {
                "type": "string"
              },
              "dataNascimento": {
                "type": "string"
              },
              "numeroTelemovel": {
                "type": "string"
              },
              "sexo": {
                "type": "string"
              }
            }
        },
...
        "Evento":{
            "type":"object",
            "properties": {
                "id":{
                    "type":"string"
                },
                "nome":{
                    "type":"string"
                },
                "conteudo":{
                    "type":"string"
                },
                "dataInicio":{
                    "type":"string"
                },
                "dataFim":{
                    "type":"string"
                }
            }
        },
...
    }   
}